+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Each tagged release MUST have a section 2 heading starting at the time of releas
 - Added the ability to override images used for all services using environment variables (T306023)
 - Added the ability to run multiple fresh (and quibble) commands simultaneously (T305683)
 - Added the `restart` command to `stop` and then `start` the current running containers (T305943)
+- Added `mysql mysql` and `mysql-replica mysql` commands that run the `mysql` cli in the mysq containers (T306864)
 - Changed `resume` to `start` with a backward compatible alias (T305823)
 - Changed `suspend` to `stop` with a backward compatible alias (T305823)
 - Improved HTML error message when MediaWiki database can not be found, including commands that might help (T305099)